requirements.txt
Flask==2.2.2
flask-sqlalchemy==3.0.2
parse==1.19.0